% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{checkKKT}
\alias{checkKKT}
\title{Check the KKT condition of lasso solutions}
\usage{
checkKKT(y, X, lambda, beta.hat, tol = 1e-05)
}
\arguments{
\item{y, }{the response vector.}

\item{X, }{the design matrix.}

\item{lambda, }{the tuning parameter.}

\item{beta.hat, }{the lasso solution.}

\item{tol, }{the tolerance.}
}
\value{
This function returns the desired logical result.
}
\description{
This function returns \code{TRUE} if and only if the lasso solution
satisfy the KKT condition.
}
\details{
This function assumes the quadratic loss is scaled by \eqn{1/2n}. That is,
    \deqn{1/(2n) * ||y-X\beta||_2^2 + \lambda * ||\beta||_1}
}
\keyword{internal}
