% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outference.R
\name{coef.outference}
\alias{coef.outference}
\title{Extract coefficients from an \code{"outference"} object}
\usage{
coef.outference(object, model = c("removed", "original"), ...)
}
\arguments{
\item{object, }{an object of class \code{"outference"}.}

\item{model, }{if \code{model = "removed"}, then the coefficients are from the fit with outliers removed;
if \code{model = "original"}, then the coefficients are from the fit with all the data.}

\item{..., }{other arguments.}
}
\value{
This function returns the desired vector of regression coefficients.
}
\description{
This function extracts the estimated regression coefficients from an \code{"outference"} object.
}
\examples{
## Brownlee’s Stack Loss Plant Data
data("stackloss")
## fit the model
## detect outlier using Cook's distance with cutoff = 4
fit <- outference(stack.loss ~ ., data = stackloss, method = "cook", cutoff = 4)
coef(fit, model = "original")    # the coefficients from the fit using all the data
coef(fit, model = "removed")     # the coefficients from the fit after outlier removal
}
\seealso{
\code{\link{outference}} for model fitting;

    \code{\link{summary.outference}} for summaries;

    \code{\link{confint.outference}} for confidence intervals of regression coefficients;

    \code{\link{plot.outference}} for plotting the outlying measure;

    \code{\link{predict.outference}} for making predictions.
}
\author{
Shuxiao Chen <sc2667@cornell.edu>
}
