% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_inf.R
\name{coefForFEachSlice}
\alias{coefForFEachSlice}
\title{Compute coeffcients that defines one slice of the truncation set for the \eqn{F} statistic}
\usage{
coefForFEachSlice(Q, a, b, wDelta, w2, z, r, tol = 1e-10)
}
\arguments{
\item{Q, }{a matrix.}

\item{a, }{a vector.}

\item{b, }{a scaler.}

\item{wDelta, w2, z, r}{coefficients related to \eqn{F} statistic.}

\item{tol, }{if the absolute value of a quantity is less than \code{tol},
then set it to \eqn{0}.}
}
\value{
This function returns a list of scaler coefficients
    \code{(x11, x12, x22, x1, x2, x0)}.
}
\description{
This function computes the coefficients \eqn{(x_11, x_12, x_22, x_1, x_2, x_0)} that defines
    one slice of the truncation set for the \eqn{F} statistic.
}
\details{
The constraint in the response \eqn{{y^T Q y + a^T y + b \ge 0}} is equivalent to
    an inequality constraint in the \eqn{F} statistic, where
    the inequality is parameterized by \eqn{(x_11, x_12, x_22, x_1, x_2, x_0)}.
}
\keyword{internal}
