% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_inf.R
\name{coefForFStatistic}
\alias{coefForFStatistic}
\title{Compute the coefficients related to truncated \eqn{F} test}
\usage{
coefForFStatistic(y, X, outlier.det, g)
}
\arguments{
\item{y, }{the response.}

\item{X, }{the design matrix (including intercepts).}

\item{outlier.det, }{indexes of outliers detected.}

\item{g, }{a subset of \eqn{[p]} indicating which group structure to test.}
}
\value{
This function returns a list of \code{(M, Mc, g, gc, Psub, Pfull, f, wDelta, w2, z, r, C)}.
    See the manuscript for details of those quantities.
}
\description{
This function computes the coefficients related to the truncated \eqn{F} test.
}
\keyword{internal}
