% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_inf.R
\name{coefForZEachSlice}
\alias{coefForZEachSlice}
\title{Compute coeffcients that defines one slice of the truncation set for the \eqn{Z} statistic}
\usage{
coefForZEachSlice(Q, a = NULL, b = NULL, v, z, sigma)
}
\arguments{
\item{Q, }{a matrix.}

\item{a, }{a vector.}

\item{b, }{a scaler.}

\item{v, }{the contrast vector \eqn{\nu}.}

\item{z, }{\eqn{P_\nu^\perp y}.}

\item{sigma, }{the noise level \eqn{\sigma}.}
}
\value{
This function returns a vector of scaler coefficients \eqn{(A, B, C)}.
}
\description{
This function computes the coefficients \eqn{(A, B, C)} that defines
    one slice of the truncation set for the \eqn{Z} statistic.
}
\details{
Consider the \eqn{Z}-statistic \deqn{Z = \nu^T y / (\sigma ||\nu||_2).} The constraint
    in the response \eqn{{y^T Q y + a^T y + b \ge 0}} is equivalent to
    \eqn{{AZ^2 + BZ + C \ge 0}} for some scaler coefficients \eqn{(A, B, C)}.
}
\keyword{internal}
