% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_inf.R
\name{compIntForChisqEachSlice}
\alias{compIntForChisqEachSlice}
\title{Solve the roots of quadratic polynomials related to truncated chi-squared test}
\usage{
compIntForChisqEachSlice(coef, tol = 1e-15)
}
\arguments{
\item{coef, }{the coefficients \eqn{(A, B, C)} of the quadratic polynomial.}

\item{tol, }{the tolerance of roots.}
}
\value{
This function returns an "Intervals" object.
}
\description{
This function solves \eqn{{x^2: Ax^2 + Bx + C \ge 0}}, then
    returns the complement of the solution set (w.r.t. to \eqn{\mathbb{R}}).
}
\details{
The reason we compute the complement instead of the original solution set
    is that when taking intersection of multiple sets, the strategy of
    "taking union of complements, then taking comlement" is substantially faster
    than taking intersections directly based on the \code{intervals} package.
}
\keyword{internal}
