% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_inf.R
\name{computeCI}
\alias{computeCI}
\title{compute selective confidence intervals}
\usage{
computeCI(v, y, sigma, truncation, alpha)
}
\arguments{
\item{v, }{the contrast vector.}

\item{y, }{the response.}

\item{sigma, }{the noise level \eqn{\sigma}.}

\item{truncation, }{the truncation set for the \eqn{Z}-statistic.}

\item{alpha, }{the significance level.}
}
\value{
This function returns a vector of lower and upper confidence limits.
}
\description{
This function computes the selective confidence intervals.
}
\keyword{internal}
