% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outference.R
\name{confint.outference}
\alias{confint.outference}
\title{Form confidence intervals for regression coeffcients from an \code{"outference"} object}
\usage{
\method{confint}{outference}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object, }{an object of class \code{"outference"}.}

\item{parm, }{indexes of which parameter to consider. If set to \code{NULL}, then
all parameters are considered.}

\item{level, }{the confidence level.}

\item{..., }{other arguments.}
}
\value{
A matrix with columns being lower and upper confidence limits for each parameter.
}
\description{
This function constructs confidence intervals for the regression coefficients from an
    \code{"outference"} object. The syntax is the same as \code{\link{confint.lm}}.
}
\examples{
## Brownlee’s Stack Loss Plant Data
data("stackloss")
## fit the model
## detect outlier using Cook's distance with cutoff = 4
fit <- outference(stack.loss ~ ., data = stackloss, method = "cook", cutoff = 4)
confint(fit)

}
\seealso{
\code{\link{outference}} for model fitting;

    \code{\link{summary.outference}} for summaries;

    \code{\link{coef.outference}} for extracting coefficients;

    \code{\link{plot.outference}} for plotting the outlying measure;

    \code{\link{predict.outference}} for making predictions.
}
\author{
Shuxiao Chen <sc2667@cornell.edu>
}
