% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_inf.R
\name{constrInResponseCook}
\alias{constrInResponseCook}
\title{Compute the truncation set in the response after
    outlier detection using Cook's distance}
\usage{
constrInResponseCook(n, p, PX, PXperp, outlier.det, cutoff)
}
\arguments{
\item{n, }{the number of observations.}

\item{p, }{the number of variables, including the intercept.}

\item{PX, }{the projection matrix onto the column space of the design matrix \eqn{X}.}

\item{PXperp, }{\code{I - PX}.}

\item{outlier.det, }{indexes of detected outliers, can be empty.}

\item{cutoff, }{the cutoff \eqn{\lambda} (see details).}
}
\value{
This function returns a list of matrices \eqn{Q_i}.
}
\description{
This function computes the matrices \eqn{Q_i}, such that
    outlier detection using Cook's distance is equivalent to
    \eqn{\bigcap_{i \in [n]} y^T Q_i y \ge 0}.
}
\details{
Using Cook's distance as a heuristic, the \eqn{i}-th data is considered
    as an outlier if and only if its Cook's distance is larger than \eqn{\lambda/n},
    where \eqn{lambda} is the user-specified cutoff. Then we can characterize this
    "detection event" as an intersection of quadratic constraints in the response \eqn{y}
    by \eqn{\bigcap_{i \in [n]} y^T Q_i y \ge 0}.
}
\keyword{internal}
