% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_inf.R
\name{constrInResponseCookSeq}
\alias{constrInResponseCookSeq}
\title{Compute the truncation set in the response after SEQUENTIAL
    outlier detection using Cook's distance}
\usage{
constrInResponseCookSeq(n, p, PX, PXperp, obs.ordered, numOfOutlier)
}
\arguments{
\item{n, }{the number of observations.}

\item{p, }{the number of variables, including the intercept.}

\item{PX, }{the projection matrix onto the column space of the design matrix \eqn{X}.}

\item{PXperp, }{\code{I - PX}.}

\item{obs.ordered, }{the index of observations with their cook's distance in the decreasing order.}

\item{numOfOutlier, }{the number of outliers assumed, must be between \eqn{0} and \eqn{n}.}
}
\value{
This function returns a list of matrices \eqn{Q_i}.
}
\description{
This function computes the matrices \eqn{Q_i}, such that sequential
    outlier detection using Cook's distance is equivalent to
    \eqn{\bigcap_{i \in I} y^T Q_i y \ge 0}.
}
\details{
Using Cook's distance sequentially and assume there are \eqn{k} outliers,
    the \eqn{i}-th data is considered
    as an outlier if and only if its Cook's distance ranks in top-\eqn{k} among all,
    Then we can characterize this
    "detection event" as an intersection of quadratic constraints in the response \eqn{y}
    by \eqn{\bigcap_{i \in I} y^T Q_i y \ge 0}, where \eqn{I} is a finite index set.
}
\keyword{internal}
