% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_inf.R
\name{constrInResponseLasso}
\alias{constrInResponseLasso}
\title{Compute the truncation set in the response after
    outlier detection using lasso}
\usage{
constrInResponseLasso(n, p, PXperp, outlier.det, outlier.det.sign, cutoff)
}
\arguments{
\item{n, }{the number of observations.}

\item{p, }{the number of variables, including the intercept.}

\item{PXperp, }{the projection matrix onto the orthogonal complement
of the column space of the design matrix \eqn{X}.}

\item{outlier.det, }{indexes of detected outliers, can be empty.}

\item{outlier.det.sign, }{the sign of the active variable estimated by lasso.}

\item{cutoff, }{the cutoff \eqn{\lambda} (see details).}
}
\value{
This function returns a list (A, b).
}
\description{
This function computes a matrix \eqn{A} and a vector \eqn{b}, such that
    outlier detection using lasso is equivalent to
    \eqn{A y \ge b}.
}
\details{
Consider solving the following program
    \deqn{minimize ||y-X\beta-u||_2^2/(2n) + \lambda ||u||_1.}
    The \eqn{i}-th observation is considered as an outlier
    if and only if \eqn{\hat u_i \neq 0}.
    This is equivalent to solving
    \deqn{minimize ||P_X^\perp (y-u)||_2^2/(2n) + \lambda ||u||_1.}
    Then the variable selection can be characterized by a set of affine constraints
    \eqn{Ay \ge b}. In essence, this function is equivalent to
    \code{selectiveInference:::fixedLasso.poly}, but adapted to our notations
    and up to some scaling factors and linear transformations.
}
\references{
Lee, Jason D., et al. "Exact post-selection inference, with application to the lasso."
    The Annals of Statistics 44.3 (2016): 907-927.

Tibshirani, R., et al. "selectiveInference: Tools for Post-Selection Inference."
    R package version 1.3 (2016).
}
\keyword{internal}
