% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_inf.R
\name{contrastForCoef}
\alias{contrastForCoef}
\title{Compute the contrast vector for inference of regression coefficients}
\usage{
contrastForCoef(j, X, outlier.det)
}
\arguments{
\item{j, }{the index of regression coefficients.}

\item{X, }{the design matrix (including the intercept).}

\item{outlier.det, }{indexes of detected outliers.}
}
\value{
This function returns the contrast vector \eqn{\nu}.
}
\description{
This function computes the contrast vector \eqn{\nu} such that
    \deqn{\nu^T y = \beta^M_j.}
}
\keyword{internal}
