% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_inf.R
\name{contrastForSurf}
\alias{contrastForSurf}
\title{Compute the contrast vector for inference of regression surfaces}
\usage{
contrastForSurf(x0, X, outlier.det)
}
\arguments{
\item{x0, }{the new data point of length \eqn{p}.}

\item{X, }{the design matrix (including the intercept).}

\item{outlier.det, }{indexes of detected outliers.}
}
\value{
This function returns the contrast vector \eqn{\nu}.
}
\description{
This function computes the contrast vector \eqn{\nu} such that
    \deqn{\nu^T y = x_0^T\beta^M,}
    where \eqn{x_0} is a \eqn{p}-dimensional new data point.
}
\keyword{internal}
