% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{estimateLambda}
\alias{estimateLambda}
\title{Estimate the tuning parameter of lasso regression}
\usage{
estimateLambda(X, sigma, nsim = 1000)
}
\arguments{
\item{X, }{the design matrix.}

\item{sigma, }{the noise level.}

\item{nsim, }{the number of Monte Carlo simulations to run.}
}
\value{
This function returns the value of tuning parameter.
}
\description{
This function returns \eqn{\lambda = E[||X^T \epsilon||_\infty] / n},
    where \eqn{\epsilon ~ N(0, \sigma^2 I)}.
}
\keyword{internal}
