% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{estimateSigma}
\alias{estimateSigma}
\title{Estimate the noise level in linear regression with presence of outliers}
\usage{
estimateSigma(y, X)
}
\arguments{
\item{y, }{the response.}

\item{X, }{the design matrix with the first column being 1 (the column of intercepts).}
}
\value{
This function returns an estimate of the noise level \eqn{\sigma}.
}
\description{
This function estimates the noise level \eqn{\sigma} in linear regression
    with possible presence of outliers, based on "lasso-refitting" strategy.
}
\details{
Assume the mean-shift model \deqn{y = X \beta + u + \epsilon,} where
    \eqn{\epsilon ~ N(0, \sigma^2 I)}. This is equivalent to
    \deqn{y = X.enlarged \beta + \epsilon,} where \eqn{X.enlarged = (X : I_n)}.
    This function fits a lasso regression based on \eqn{(y, X.enlarged)}
    with the cross-validated tuning parameter,
    and then computes the residual sum of square, scaled by \eqn{1/(n-s)},
    where \eqn{s} is the number of active variables estimated by lasso.
}
\keyword{internal}
