% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{giveTwoSidePval}
\alias{giveTwoSidePval}
\title{Give two-sided p-values}
\usage{
giveTwoSidePval(surv)
}
\arguments{
\item{surv, }{usually the one-sided p-value.}
}
\value{
This function returns the desired two-sided p-value.
}
\description{
This function returns \code{2 * min(surv, 1-surv)}
}
\keyword{internal}
