% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outference.R
\name{grptest}
\alias{grptest}
\title{Test group structures from an \code{"outference"} object}
\usage{
grptest(object, group)
}
\arguments{
\item{object, }{an object of class \code{"outference"}.}

\item{group, }{a vector indicating which group to test. The intercept is labelled as \code{1}.}
}
\value{
This function returns a list with the following
    components:
    \item{statistic, }{the value of the chi-squared statistic or \eqn{F}-statistic,
        depending on whether \code{sigma} is known}
    \item{df, }{the degree(s) of freedom.}
    \item{truncation, }{the truncation set for the statistic.}
    \item{pval, }{the corrected p-value.}
}
\description{
This function does a selective test for group structures
    (test whether the regression coefficients indexed by \code{g} is all zero or not)
    from an \code{"outference"} object.
}
\examples{
## Brownlee’s Stack Loss Plant Data
data("stackloss")
## fit the model
## detect outlier using Cook's distance with cutoff = 4
fit <- outference(stack.loss ~ ., data = stackloss, method = "cook", cutoff = 4)
grptest(fit, 2:4)

}
\seealso{
\code{\link{coeftest}} for testing for a single regression coefficient
    from an \code{"outference"} object.
}
\author{
Shuxiao Chen <sc2667@cornell.edu>
}
