% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_inf.R
\name{intForChisqAll}
\alias{intForChisqAll}
\title{Compute the truncation set for chi-squared statistic}
\usage{
intForChisqAll(method = c("cook", "dffits", "lasso"), constraint, w, z, sigma,
  outlier.det)
}
\arguments{
\item{method, }{the outlier detection method, must be one of "cook", "dffits", "lasso".}

\item{constraint, }{the constraint in the response.}

\item{w, z, }{coefficiens related to chi-squared statistic}

\item{sigma, }{the noise level \eqn{\sigma}.}

\item{outlier.det, }{indexes of detected outliers.}
}
\value{
This function returns an "Intervals" object.
}
\description{
This function computes the truncation set for the chi-squared statistc.
}
\details{
This function translates the constraints in the response into the truncation
    set (which is a union of intervals) for the chi-squared statistic.
}
\keyword{internal}
