% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_inf.R
\name{intForFAll}
\alias{intForFAll}
\title{Compute the truncation set for the \eqn{F} statistic}
\usage{
intForFAll(method = c("cook", "dffits", "lasso"), constraint, wDelta, w2, z,
  r, C, outlier.det)
}
\arguments{
\item{method, }{the outlier detection method, must be one of "cook", "dffits", "lasso".}

\item{constraint, }{the constraint in the response.}

\item{wDelta, w2, z, r, C, }{coefficiens related to the \eqn{F} statistic.}

\item{outlier.det, }{indexes of detected outliers.}
}
\value{
This function returns an "Intervals" object.
}
\description{
This function computes the truncation set for the \eqn{F} statistc.
}
\details{
This function translates the constraints in the response into the truncation
    set (which is a union of intervals) for the \eqn{F} statistic.
}
\keyword{internal}
