% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_inf.R
\name{intForZAll}
\alias{intForZAll}
\title{Compute the truncation set for \eqn{Z} statistic}
\usage{
intForZAll(method = c("cook", "dffits", "lasso"), constraint, v, z, sigma,
  outlier.det)
}
\arguments{
\item{method, }{the outlier detection method, must be one of "cook", "dffits", "lasso".}

\item{constraint, }{the constraint in the response.}

\item{v, }{the contrast vector \eqn{\nu}.}

\item{z, }{\eqn{P_\nu^\perp y}.}

\item{sigma, }{the noise level \eqn{\sigma}.}

\item{outlier.det, }{indexes of detected outliers.}
}
\value{
This function returns an "Intervals" object.
}
\description{
This function computes the truncation set for the \eqn{Z} statistc.
}
\details{
Consider the \eqn{Z}-statistic \deqn{Z = \nu^T y / (\sigma ||\nu||_2).}
    This function translates the constraints in the response into the truncation
    set (which is a union of intervals) for the \eqn{Z} statistic.
}
\keyword{internal}
