% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outference.R
\name{outference}
\alias{outference}
\alias{print.outference}
\alias{print.outference}
\title{Fit a linear model with outliers detected and removed}
\usage{
outference(formula, data, method = c("cook", "dffits", "lasso"),
  cutoff = NULL, sigma = NULL)

\method{print}{outference}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{formula, }{an object of class \code{"\link{formula}"}, the same syntax as in \code{\link{lm}}.}

\item{data, }{an optional data frame, list or environment containing the variables in the model, the same
syntax as in \code{\link{lm}}.}

\item{method, }{the outlier detection method, must be one of \code{"cook", "dffits", "lasso"}. See also 'details'.}

\item{cutoff, }{the cutoff of the outlier detection method. If \code{cutoff = NULL}, then this function
uses the default values. For \code{"cook"} or \code{"dffits"}, the default cutoff is \eqn{4};
for \code{"lasso"}, the default cutoff is \eqn{0.75*E[||X^T \epsilon||_\infty]/n}.
See also 'details'.}

\item{sigma, }{the noise level. Must be one of \code{NULL, "estimate"}, or a positive scaler value. If
\code{sigma = NULL}, then the inference will assume the noise level is unknown;
if \code{sigma = "estimate"}, then the inference will base on an estimated noise level.}

\item{x, }{an object of class \code{"outference"}.}

\item{digits, }{the number of significant digits to use when printing.}

\item{..., }{other arguments.}
}
\value{
This function returns an object of \code{\link{class}} \code{"outference"}.

    The function \code{\link{summary}} is used to obtain and print a summary (including p-values)
    of the results. The generic functions \code{\link{coef}}, \code{\link{confint}}, \code{\link{plot}},
    \code{\link{predict}} are used to extract useful features of the object returned by this function.

    An object of class \code{"outference"} is a list containing the following components:


    \item{fit.full, }{an \code{"lm"} object representing the fit using the full data (no outliers are removed).}
    \item{fit.rm, }{an \code{"lm"} object representing the fit using the data after outlier removal.}
    \item{method, }{the method used for outlier detection.}
    \item{cutoff, }{the cutoff of the method.}
    \item{outlier.det, }{indexes of detected outliers.}
    \item{magnitude, }{a measure of "outlying-ness". For \code{"cook"} and \code{"dffits"}, this is
        the vector of the Cook's distance or DFFITS for all observations; for \code{"lasso"}, this is
        the vector of "outlying coefficients" estimated by lasso. See also 'details'.}
    \item{constraint, }{the constraint in the response that characterizes the outlier detection event.
        For \code{"cook"} and \code{"dffits"}, this is a list of \eqn{n} by \eqn{n} matrices;
        for \code{"lasso"}, this is a list of \code{(A, b)}, where \code{A} is a matrix and
        \code{b} is a vector. See also 'details'.}
    \item{sigma, }{the noise level used in the fit.}
    \item{call, }{the function call.}
}
\description{
This function detects outliers using user-specified method, and
    fits a linear regression model with outliers removed. The object
    returned by this function can be used for valid inference corrected
    for outlier removal through generic functions like \code{\link{summary}},
    \code{\link{confint}}, \code{\link{predict}}.
}
\details{
This function uses the same syntax as \code{\link{lm}} for the \code{formula} and \code{data} arguments.
    Users can access the original \code{"lm"} objects through \code{$fit.full} and \code{$fit.rm}.
    Common generic functions for \code{lm}, including \code{\link{coef}}, \code{\link{confint}},
    \code{\link{plot}}, \code{\link{predict}} and \code{\link{summary}} are re-written so that
    they can be used to extract useful features of the object returned by this function.

Currently, this function supports three outlier detection methods. For \code{"cook"}, the \eqn{i}-th
    observation is considered as an outlier when its Cook's distance is greater than \code{cutoff/n},
    where \code{n} is the number of observations. For \code{"dffits"}, the \eqn{i}-th observation is
    considered as an outlier when the square of its DFFITS measure is greater than \code{cutoff*p/(n-p)},
    where \code{p} is the number of variables (including the intercept). The rule of thumb of \code{cutoff}
    for both methods are 4, which is the default value when one sets \code{cutoff = NULL}.
    The outlier detection event of both methods can be characterized as a set of quadratic constraints
    in the response \eqn{y}:
    \deqn{\bigcap_{i \in [n]} {y^T Q_i y \ge 0},}
    and the constraint returned by this function is the list of \eqn{Q_i} matrices.
    For \code{"lasso"}, we assume the \emph{mean-shift model}
    \eqn{y = X \beta + u + \epsilon}, where \eqn{u} is the "outlying coefficients" and
    \eqn{\epsilon ~ N(0, \sigma^2 I)} is the noise. We solve the following program:
    \deqn{(\hat \beta, \hat u) = argmin ||y-X\beta-u||_2^2 + cutoff*||u||_1.} The \eqn{i}-th observation
    is considered as an outlier when \eqn{\hat u_i} differs from \eqn{0}. The default cutoff for
    \code{"lasso"} is \eqn{0.75*E[||X^T \epsilon||_\infty]/n}, which is a less conservative choice
    than the prediction-optimal cutoff \eqn{2*E[||X^T \epsilon||_\infty]/n}. This cutoff is computed
    by Monte Carlo simulation and \eqn{\sigma} is replaced by an estimate when the true noise level
    is unknown. The outlier detection event of \code{"lasso"} can be characterized as a
    set of affine constraints in the response \eqn{y}:
    \deqn{A y \ge b, }
    where the \eqn{"\ge"} is interpreted as element-wise. The constraint returned by this function is
    then a list of \code{(A, b)}.
}
\examples{
## Brownlee’s Stack Loss Plant Data
data("stackloss")
head("stackloss")     # look at the dataset
## fit the model
## detect outlier using Cook's distance with cutoff = 4
fit <- outference(stack.loss ~ ., data = stackloss, method = "cook", cutoff = 4)
plot(fit)             # plot the Cook's distance of each observation
## observation 21 is considered as an outlier with cutoff = 4
summary(fit$fit.full) # look at the fit with all the data
summary(fit$fit.rm)   # look at the fit with observation 21 deleted
summary(fit)          # extract the corrected p-values after outlier removal
}
\references{
Lee, Jason D., et al. "Exact post-selection inference, with application to the lasso."
    The Annals of Statistics 44.3 (2016): 907-927.

OUR PAPER.
}
\seealso{
\code{\link{summary.outference}} for summaries;

    \code{\link{coef.outference}} for extracting coefficients;

    \code{\link{confint.outference}} for confidence intervals of regression coefficients;

    \code{\link{plot.outference}} for plotting the outlying measure;

    \code{\link{predict.outference}} for making predictions.
}
\author{
Shuxiao Chen <sc2667@cornell.edu>
}
