% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outference.R
\name{outference_seq}
\alias{outference_seq}
\title{Fit a linear model with outliers detected SEQUENTIALLY}
\usage{
outference_seq(formula, data, sigma = NULL, numOfOutlier)
}
\arguments{
\item{formula, }{an object of class \code{"\link{formula}"}, the same syntax as in \code{\link{lm}}.}

\item{data, }{an optional data frame, list or environment containing the variables in the model, the same
syntax as in \code{\link{lm}}.}

\item{sigma, }{the noise level. Must be one of \code{NULL, "estimate"}, or a positive scaler value. If
\code{sigma = NULL}, then the inference will assume the noise level is unknown;
if \code{sigma = "estimate"}, then the inference will base on an estimated noise level.}

\item{numOfOutlier, }{the number of outliers to be detected.}
}
\value{
This function returns an object of \code{\link{class}} \code{c("outference_seq", "outference")}.

    The function \code{\link{summary}} is used to obtain and print a summary (including p-values)
    of the results. The generic functions \code{\link{coef}}, \code{\link{confint}}, \code{\link{plot}},
    \code{\link{predict}} are used to extract useful features of the object returned by this function.

    An object of class \code{c("outference_seq", "outference")} is a list containing the following components:

    \item{fit.full, }{an \code{"lm"} object representing the fit using the full data (no outliers are removed).}
    \item{fit.rm, }{an \code{"lm"} object representing the fit using the data after outlier removal.}
    \item{method, }{"cook".}
    \item{cutoff, }{\code{NULL}.}
    \item{numOfOutlier, }{the number of outliers to be detected.}
    \item{outlier.det, }{indexes of detected outliers.}
    \item{magnitude, }{ the vector of the Cook's distance for all observations}
    \item{constraint, }{the constraint in the response that characterizes the outlier detection event.
        A list of \eqn{n} by \eqn{n} matrices.}
    \item{sigma, }{the noise level used in the fit.}
    \item{call, }{the function call.}
}
\description{
This function detects outliers by using Cook's distance sequentially, and
    fits a linear regression model with outliers removed. The object
    returned by this function can be used for valid inference corrected
    for outlier removal through generic functions like \code{\link{summary}},
    \code{\link{confint}}, \code{\link{predict}}.
}
\details{
This function uses the same syntax as \code{\link{lm}} for the \code{formula} and \code{data} arguments.
    Users can access the original \code{"lm"} objects through \code{$fit.full} and \code{$fit.rm}.
    Common generic functions for \code{lm}, including \code{\link{coef}}, \code{\link{confint}},
    \code{\link{plot}}, \code{\link{predict}} and \code{\link{summary}} are re-written so that
    they can be used to extract useful features of the object returned by this function.

The \eqn{i}-th observation is considered as an outlier when its Cook's distance
    rank among top \eqn{k}, where \eqn{k} is the user-specified number of outliers to be detected.
    The outlier detection event can be characterized as a set of quadratic constraints
    in the response \eqn{y}:
    \deqn{\bigcap_{i \in I} {y^T Q_i y \ge 0},}
    where \eqn{I} is a finite index set, and the constraint returned by this function
    is the list of \eqn{Q_i} matrices.
}
\references{
S. Chen and J. Bien. “Valid Inference Corrected for Outlier Removal”. arXiv preprint arXiv:1711.10635 (2017).
}
\seealso{
\code{\link{summary.outference}} for summaries;

    \code{\link{coef.outference}} for extracting coefficients;

    \code{\link{confint.outference}} for confidence intervals of regression coefficients;

    \code{\link{plot.outference}} for plotting the outlying measure;

    \code{\link{predict.outference}} for making predictions.
}
\author{
Shuxiao Chen <sc2667@cornell.edu>
}
\keyword{internal}
