% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outference.R
\name{plot.outference}
\alias{plot.outference}
\title{Plot the outlying measure from an \code{"outference"} object}
\usage{
\method{plot}{outference}(x, ...)
}
\arguments{
\item{x, }{an object of class \code{"outference"}.}

\item{..., }{other arguments.}
}
\description{
This function plots the Cook's distance, or DFFITS, or the estimated
outlying coefficients, as well as the cutoff based on the outlier detection method.
}
\examples{
## Brownlee’s Stack Loss Plant Data
data("stackloss")
## fit the model
## detect outlier using Cook's distance with cutoff = 4
fit <- outference(stack.loss ~ ., data = stackloss, method = "cook", cutoff = 4)
plot(fit)             # plot the Cook's distance of each observation

}
\seealso{
\code{\link{outference}} for model fitting;

    \code{\link{summary.outference}} for summaries;

    \code{\link{coef.outference}} for extracting coefficients;

    \code{\link{confint.outference}} for confidence intervals of regression coefficients;

    \code{\link{predict.outference}} for making predictions.
}
\author{
Shuxiao Chen <sc2667@cornell.edu>
}
