% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outference.R
\name{predict.outference}
\alias{predict.outference}
\title{Make predictions from an \code{"outference"} object}
\usage{
\method{predict}{outference}(object, newdata, interval = c("none",
  "confidence", "prediction"), level = 0.95, alpha.tilde = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{"outference"}.}

\item{newdata, }{an optional data frame in which to look for variables with which to predict. If
omitted, the fitted values are used. WARNING: making predictions for many new data points with
\code{interval = "confidence"} or \code{"prediction"} can be time-consuming,
since for each data point, the function needs to compute the truncation set and
solve the roots for a truncated survival function.}

\item{interval, }{type of interval calculation. If set to \code{"none"}, then only
point predictions are made; if set to \code{"confidence"}, then this function returns
confidence intervals for the regression surface; if set to \code{"prediction"}, then this
function returns the prediction intervals.}

\item{level, }{confidence level, default to \eqn{0.95}.}

\item{alpha.tilde, }{an extra parameter between \code{0} and \code{1-level}, which is  used when
computing prediction intervals. If left \code{NULL}, then this function searches the
\code{alpha.tilde} that gives shortest prediction intervals. See also 'details'.}

\item{..., }{other arguments.}
}
\value{
This function gives a vector of predictions or a matrix of predictions and intervals
    with column names \code{fit, lwr, upr} if \code{interval} is set.
}
\description{
This function gives predictions as well as confidence intervals for the regression surfaces
    and the prediction intervals from an \code{"outference"} object. The syntax is the same
    as \code{\link{predict.lm}}.
}
\details{
If \code{alpha.tilde = NULL}, then this function iterates over
    \code{alpha.tilde in seq(0, 1-level, length.out = 100)[c(5, 25, 50, 75, 95)]} and returns
    the results with the shortest prediction intervals.
}
\examples{
## Brownlee’s Stack Loss Plant Data
data("stackloss")
## fit the model
## detect outlier using Cook's distance with cutoff = 4
fit <- outference(stack.loss ~ ., data = stackloss, method = "cook", cutoff = 4)
## predictions at the first two observations
predict(fit, newdata = stackloss[1:2, ], interval = "none")
predict(fit, newdata = stackloss[1:2, ], interval = "confidence")
predict(fit, newdata = stackloss[1:2, ], interval = "prediction")

}
\seealso{
\code{\link{outference}} for model fitting;

    \code{\link{summary.outference}} for summaries;

    \code{\link{coef.outference}} for extracting coefficients;

    \code{\link{confint.outference}} for confidence intervals of regression coefficients;

    \code{\link{plot.outference}} for plotting the outlying measure;
}
\author{
Shuxiao Chen <sc2667@cornell.edu>
}
