% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{proj}
\alias{proj}
\title{Projection matrix onto the column space}
\usage{
proj(X)
}
\arguments{
\item{X, }{a matrix with full column rank, can be an empty matrix with \code{ncol(X) == 0}.}
}
\value{
This function returns the desired square matrix with dimension \code{nrow(X)}.
}
\description{
This function assumes \eqn{X} has full column rank, and
    returns the projection matrix onto the column space of \eqn{X}.
}
\details{
The projection matrix onto the column space of \eqn{X} can be written as
    \deqn{X(X^TX)^{-1}X^T.}
    When \eqn{X} is an empty matrix, then this function returns the zero matrix.
}
\keyword{internal}
