% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outference.R
\name{summary.outference}
\alias{summary.outference}
\alias{print.summary.outference}
\alias{print.summary.outference}
\title{Summarize from an \code{"outference"} object}
\usage{
summary.outference(object, ...)

print.summary.outference(x, digits = max(3, getOption("digits") - 3),
  signif.stars = getOption("show.signif.stars"), ...)
}
\arguments{
\item{object, }{an object of class \code{"outference"}.}

\item{..., }{other arguments.}

\item{x, }{an object of class \code{"summary.outference"}.}

\item{digits, }{the number of significant digits to use when printing.}

\item{signif.stars, }{should the 'significance starts' be printed?}
}
\value{
This function returns an object of class \code{"summary.outference"}, which is a list containing
    the following components:
    \item{call, }{the function call.}
    \item{summary.full, }{an object of class \code{"summary.lm"}, representing the summary
        from the fit using the full data.}
    \item{summary.rm, }{an object of class \code{"summary.lm"}, representing the summary
        from the fit after outlier removal.}
    \item{method, }{the method used for outlier detection.}
    \item{cutoff, }{the cutoff of the method.}
    \item{outlier.det, }{indexes of detected outliers.}
    \item{magnitude, }{a measure of "outlying-ness". For \code{"cook"} and \code{"dffits"}, this is
        the vector of the Cook's distance or DFFITS for all observations; for \code{"lasso"}, this is
        the vector of "outlying coefficients" estimated by lasso.}
    \item{sigma, }{the noise level used in the fit.}
    \item{coefficients, }{a data frame summarizing the estimates, standard errors, values of the test
        statistics and corrected p-values of regression coefficients.}
    \item{truncation.coef, }{a list of the truncation sets of the test statistics for each regression
        coefficient.}
    \item{chisqstatistic, fstatistic, }{a list containing the value, the degree(s) of freedom, the
        truncation set, and the corrected p-value for testing the global null.}
}
\description{
This function produces a summary from an \code{"outference"} object, with a similar fasion
    as \code{\link{summary.lm}}.
}
\details{
This function is written in a similar fasion as \code{\link{summary.lm}}. Users can get access
    to the \code{"summary.lm"} objects through \code{$summary.full} and \code{$summary.rm}.
}
\examples{
## Brownlee’s Stack Loss Plant Data
data("stackloss")
## fit the model
## detect outlier using Cook's distance with cutoff = 4
fit <- outference(stack.loss ~ ., data = stackloss, method = "cook", cutoff = 4)
summary(fit)          # extract the corrected p-values after outlier removal
}
\seealso{
\code{\link{outference}} for model fitting;

    \code{\link{coef.outference}} for extracting coefficients;

    \code{\link{confint.outference}} for confidence intervals of regression coefficients;

    \code{\link{plot.outference}} for plotting the outlying measure;

    \code{\link{predict.outference}} for making predictions.
}
\author{
Shuxiao Chen <sc2667@cornell.edu>
}
