% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_distr.R
\name{TChisqProb}
\alias{TChisqProb}
\title{Probability of a chi-squared random variable in a union of intervals}
\usage{
TChisqProb(df, E)
}
\arguments{
\item{df}{degree of freedom of the chi-squared random variable.}

\item{E}{an "Intervals" object or a matrix where rows represents
a union of disjoint intervals.}
}
\value{
This function returns the desired probability.
}
\description{
This function returns \eqn{P(X \in E)}, where \eqn{X} is a
    central chi-squared random variable with \code{df} degrees of freedom.
}
\keyword{internal}
