% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_distr.R
\name{TChisqProbEachInt}
\alias{TChisqProbEachInt}
\title{Probability of a chi-squared random variable in a single interval}
\usage{
TChisqProbEachInt(df, lo, up)
}
\arguments{
\item{df}{degree of freedom of the chi-squared random variable.}

\item{lo, up}{quantiles.}
}
\value{
This function returns the desired probability.
}
\description{
This function returns \eqn{P(lo \le X \le up)}, where \eqn{X} is a
    central chi-squared random variable with \code{df} degrees of freedom.
}
\keyword{internal}
