% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_distr.R
\name{TChisqSurv}
\alias{TChisqSurv}
\title{Survival function of truncated central chi-squared distribution}
\usage{
TChisqSurv(q, df, E, approx = FALSE)
}
\arguments{
\item{q}{the quantile.}

\item{df}{the degree of freedom.}

\item{E}{the truncation set, an "Intervals" object or a matrix where rows represents
a union of disjoint intervals.}

\item{approx}{should the approximation algorithm be used? Default is \code{FALSE},
where the approximation is not used in the first place. But when the result is wacky,
the approximation will be used.}
}
\value{
This function returns the value of the survival function evaluated at quantile \code{q}.
}
\description{
This function returns the upper tail probability of a truncated central chi-squared distribution
    at quantile \code{q}.
}
\details{
Let \eqn{X} be a central chi-squared random variable with \code{df} degrees of freedom. Truncating
    \eqn{X} to the set \eqn{E} is equivalent to conditioning on \eqn{{X \in E}}. So this function
    returns \eqn{P(X \ge q | X \in E)}.
}
\references{
Bryc, Wlodzimierz. "A uniform approximation to the right normal tail integral."
    Applied mathematics and computation 127.2 (2002): 365-374.

Canal, Luisa. "A normal approximation for the chi-square distribution."
    Computational statistics & data analysis 48.4 (2005): 803-808.
}
\keyword{internal}
