% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_distr.R
\name{TFProb}
\alias{TFProb}
\title{Probability of an F random variable in a union of intervals}
\usage{
TFProb(df1, df2, E)
}
\arguments{
\item{df1, df2}{degrees of freedom of the central F random variable.}

\item{E}{an "Intervals" object or a matrix where rows represents
a union of disjoint intervals.}
}
\value{
This function returns the desired probability.
}
\description{
This function returns \eqn{P(X \in E)}, where \eqn{X} is a
    central F random variable with \code{df1, df2} degrees of freedom.
}
\keyword{internal}
