% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_distr.R
\name{TFProbEachInt}
\alias{TFProbEachInt}
\title{Probability of an F random variable in a single interval}
\usage{
TFProbEachInt(df1, df2, lo, up)
}
\arguments{
\item{df1, df2}{degrees of freedom of the central F random variable.}

\item{lo, up}{quantiles.}
}
\value{
This function returns the desired probability.
}
\description{
This function returns \eqn{P(lo \le X \le up)}, where \eqn{X} is a
    central F random variable with \code{df1, df2} degrees of freedom.
}
\keyword{internal}
