% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_distr.R
\name{TFRatioApprox}
\alias{TFRatioApprox}
\title{Approximation of the ratio of two F probabilities}
\usage{
TFRatioApprox(df1, df2, E1, E2)
}
\arguments{
\item{df1, df2}{degrees of freedom of the F random variable.}

\item{E1, E2}{"Intervals" objects or matrices where rows represents
a union of intervals with \emph{positive and finite} endpoints.}
}
\value{
This function returns the value of the approximation.
}
\description{
This function returns an approximation of \eqn{P(X \in E1)/P(X \in E2)}, where
    \eqn{X} is a central F random variable with \code{df1, df2} degrees of freedom.
}
\references{
Bryc, Wlodzimierz. "A uniform approximation to the right normal tail integral."
    Applied mathematics and computation 127.2 (2002): 365-374.

Peizer, David B., and John W. Pratt. "A normal approximation for binomial,
    F, beta, and other common, related tail probabilities, I." Journal of the American
    Statistical Association 63.324 (1968): 1416-1456.
}
\keyword{internal}
