% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_distr.R
\name{TFSurv}
\alias{TFSurv}
\title{Survival function of truncated central F distribution}
\usage{
TFSurv(q, df1, df2, E, approx = FALSE)
}
\arguments{
\item{q}{the quantile.}

\item{df1, df2}{the degrees of freedom.}

\item{E}{the truncation set, an "Intervals" object or a matrix where rows represents
a union of disjoint intervals.}

\item{approx}{should the approximation algorithm be used? Default is \code{FALSE},
where the approximation is not used in the first place. But when the result is wacky,
the approximation will be used.}
}
\value{
This function returns the value of the survival function evaluated at quantile \code{q}.
}
\description{
This function returns the upper tail probability of a truncated central F distribution
    at quantile \code{q}.
}
\details{
Let \eqn{X} be a central F random variable with \code{df1, df2} degrees of freedom. Truncating
    \eqn{X} to the set \eqn{E} is equivalent to conditioning on \eqn{{X \in E}}. So this function
    returns \eqn{P(X \ge q | X \in E)}.
}
\references{
Bryc, Wlodzimierz. "A uniform approximation to the right normal tail integral."
    Applied mathematics and computation 127.2 (2002): 365-374.

Peizer, David B., and John W. Pratt. "A normal approximation for binomial,
    F, beta, and other common, related tail probabilities, I." Journal of the American
    Statistical Association 63.324 (1968): 1416-1456.
}
\keyword{internal}
